+++ packages/fields/CHANGELOG.md
@@ -1,3 +1,5 @@
+# @keystone-alpha/fields
+
 # @voussoir/fields
 
 ## 3.1.0